+++ resources/views/administration/users/users_index.blade.php
@@ -41,9 +41,11 @@
                         </td>
                         <td>
                             <div class="acciones" >
-                                <a href="/administracion/usuarios/{{$user->id}}/edit" class="btn btn-sm btn-success">Editar</a> |
-                                <a class="btn btn-sm btn-danger"
-                                   onclick="deleteConfirm('{{$user->full_name}}','{{$user->id}}','/administracion/usuarios/')">Eliminar</a>
+                                <a href="/administracion/usuarios/{{$user->id}}/edit" class="btn btn-sm btn-success">Editar</a>
+                                @if(!$user->hasRole('administrador-del-sistema'))
+                                    | <a class="btn btn-sm btn-danger"
+                                       onclick="deleteConfirm('{{$user->full_name}}','{{$user->id}}','/administracion/usuarios/')">Eliminar</a>
+                                @endif
                             </div>
                         </td>
                     </tr>
